% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{check_margins_dependencies}
\alias{check_margins_dependencies}
\title{Check margins dependencies}
\usage{
check_margins_dependencies(margins)
}
\arguments{
\item{margins}{Logical indicating if marginal effects are requested}
}
\value{
Invisible TRUE if dependencies available, otherwise stops with error
}
\description{
Check margins dependencies
}
\keyword{internal}
