% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easytable-package.R
\docType{package}
\name{easytable-package}
\alias{easytable-package}
\title{easytable: Create Multi-Format Regression Tables}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Create publication-ready regression tables in multiple formats (Word, Markdown, LaTeX/PDF) from statistical models. Supports lm and glm models with options for robust standard errors, marginal effects, and control variable grouping. Tables can be exported to Word via flextable, Markdown for Quarto/RMarkdown documents, or LaTeX for PDF output.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/alfredo-hs/easytable}
  \item Report bugs at \url{https://github.com/alfredo-hs/easytable/issues}
}

}
\author{
\strong{Maintainer}: Alfredo Hernandez Sanchez \email{alhdzsz@gmail.com}

}
\keyword{internal}
