% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_measure_names}
\alias{get_measure_names}
\title{Get measure row names}
\usage{
get_measure_names()
}
\value{
Character vector of measure names
}
\description{
Returns the standard names used for model fit measures
}
\keyword{internal}
