% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_measure_row_indices}
\alias{get_measure_row_indices}
\title{Get indices of model statistic rows}
\usage{
get_measure_row_indices(table)
}
\description{
Get indices of model statistic rows
}
\keyword{internal}
