% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_package_available}
\alias{is_package_available}
\title{Check if a package is available}
\usage{
is_package_available(pkg)
}
\arguments{
\item{pkg}{Character string with package name}
}
\value{
Logical indicating if package is installed
}
\description{
Check if a package is available
}
\keyword{internal}
