% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{is_supported_model}
\alias{is_supported_model}
\title{Check if model is supported type}
\usage{
is_supported_model(model)
}
\arguments{
\item{model}{A statistical model object}
}
\value{
Logical indicating if model type is supported
}
\description{
Check if model is supported type
}
\keyword{internal}
