% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_unique_labels}
\alias{make_unique_labels}
\title{Ensure labels are unique via numeric suffix}
\usage{
make_unique_labels(labels)
}
\description{
Ensure labels are unique via numeric suffix
}
\keyword{internal}
