% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_models.R
\name{parse_models}
\alias{parse_models}
\title{Parse multiple models into a combined table}
\usage{
parse_models(model_list, robust.se = FALSE, margins = FALSE, digits = 2)
}
\arguments{
\item{model_list}{A named list of statistical models}

\item{robust.se}{Logical indicating whether to use robust standard errors}

\item{margins}{Logical indicating whether to compute marginal effects}

\item{digits}{Integer number of digits after the decimal point for
coefficients and standard errors, including the mantissa in scientific
notation. Allowed values are 0 to 4.}
}
\value{
A data frame with terms in rows and models in columns
}
\description{
Parse multiple models into a combined table
}
\keyword{internal}
