% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe Operator}
\usage{
lhs \%>\% rhs
}
\value{
The result of evaluating the right-hand side expression after the
left-hand side has been passed through the pipe.
}
\description{
Re-export of \verb{\%>\%} from 'magrittr'.
}
\keyword{internal}
