% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_models.R
\name{tidy_margins}
\alias{tidy_margins}
\title{Extract tidy marginal effects from a margins object}
\usage{
tidy_margins(model)
}
\arguments{
\item{model}{A statistical model object (lm or glm)}
}
\value{
A data frame with columns: term, estimate, std.error, p.value
}
\description{
Extract tidy marginal effects from a margins object
}
\keyword{internal}
