% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validate_digits}
\alias{validate_digits}
\title{Validate digits parameter}
\usage{
validate_digits(digits)
}
\arguments{
\item{digits}{Integer number of digits after the decimal point. Allowed
values are 0 to 4.}
}
\value{
Invisible TRUE if valid, otherwise stops with error
}
\description{
Validate digits parameter
}
\keyword{internal}
