% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validate_model_types}
\alias{validate_model_types}
\title{Validate all models in list are supported types}
\usage{
validate_model_types(model_list)
}
\arguments{
\item{model_list}{A named list of models}
}
\value{
Invisible TRUE if valid, otherwise stops with error
}
\description{
Validate all models in list are supported types
}
\keyword{internal}
