% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validate_output_format}
\alias{validate_output_format}
\title{Validate output format}
\usage{
validate_output_format(output)
}
\arguments{
\item{output}{Character string specifying output format}
}
\value{
A normalized output string
}
\description{
Validate output format
}
\keyword{internal}
