% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_models.R
\name{extract_model_measures}
\alias{extract_model_measures}
\title{Extract goodness-of-fit measures from a model}
\usage{
extract_model_measures(model)
}
\arguments{
\item{model}{A statistical model object (lm or glm)}
}
\value{
A data frame with model fit statistics
}
\description{
Extract goodness-of-fit measures from a model
}
\keyword{internal}
