% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_term_labels}
\alias{format_term_labels}
\title{Format term labels for display}
\usage{
format_term_labels(terms, levels_map = NULL, abbreviate = FALSE)
}
\arguments{
\item{abbreviate}{Logical. Apply abbreviation? Default FALSE}
}
\description{
Transformations:
\itemize{
\item Preserve polynomial suffixes (.L, .Q, .C, etc.)
\item Split factor levels using levels_map
\item Abbreviate only variable portion (if abbreviate = TRUE)
\item Ensure uniqueness after abbreviation
}
}
\keyword{internal}
