% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_factor_level}
\alias{split_factor_level}
\title{Detect and split factor level suffix using model metadata}
\usage{
split_factor_level(term, levels_map = NULL)
}
\description{
Detect and split factor level suffix using model metadata
}
\keyword{internal}
