% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wrap_interaction_terms}
\alias{wrap_interaction_terms}
\title{Wrap interaction terms to reduce term-column width}
\usage{
wrap_interaction_terms(terms)
}
\description{
Example: \code{"x1 * x2" -> "x1 *\\\\nx2"}
}
\keyword{internal}
